requirements.txt
altair
numpy
pandas
pydeck
streamlit
requests 
openai 
bs4 
googlesearch-python
httpx-oauth
tiktoken
google-cloud-firestore
openai-whisper
setuptools-rust
yfinance
matplotlib